+++ src/components/home-content-section/index.ts
@@ -1,3 +1,0 @@
-export * from './HorizontalContentArticle'
-export * from './VerticalContentArticle'
-export * from './blurbs'